rename hexstring-utils to unsafe-hexstring-utils
