+++ shared/modules/unsafe-hexstring-utils.js
