+++ ui/components/app/transaction-list-item-details/transaction-list-item-details.container.js
@@ -4,7 +4,7 @@ import {
   getAddressBook,
   getRpcPrefsForCurrentProvider,
 } from '../../../selectors';
-import { toChecksumHexAddress } from '../../../../shared/modules/hexstring-utils';
+import { toChecksumHexAddress } from '../../../../shared/modules/unsafe-hexstring-utils';
 import TransactionListItemDetails from './transaction-list-item-details.component';
 
 const mapStateToProps = (state, ownProps) => {
