+++ ui/components/ui/sender-to-recipient/sender-to-recipient.component.js
@@ -7,7 +7,7 @@ import Identicon from '../identicon';
 import { shortenAddress } from '../../../helpers/utils/util';
 import AccountMismatchWarning from '../account-mismatch-warning/account-mismatch-warning.component';
 import { useI18nContext } from '../../../hooks/useI18nContext';
-import { toChecksumHexAddress } from '../../../../shared/modules/hexstring-utils';
+import { toChecksumHexAddress } from '../../../../shared/modules/unsafe-hexstring-utils';
 import {
   DEFAULT_VARIANT,
   CARDS_VARIANT,
