+++ ui/selectors/selectors.js
@@ -22,7 +22,7 @@ import {
 
 import { TEMPLATED_CONFIRMATION_MESSAGE_TYPES } from '../pages/confirmation/templates';
 
-import { toChecksumHexAddress } from '../../shared/modules/hexstring-utils';
+import { toChecksumHexAddress } from '../../shared/modules/unsafe-hexstring-utils';
 import { DAY } from '../../shared/constants/time';
 import {
   getNativeCurrency,
